requirements.txt
requests
bs4
deep_translator
googlesearch-python
